feat: initial commit
This commits moves the project from `cmackenzie1/holster` to its
own repository to make it easier setup and discover. It also makes
some significant changes (i.e, deletes a bunch of boiler plate)
with the latest addition of Workers RPC. I also switched it from
`itty-router` to `hono` because I wanted to.